move HttpContext to separate helper class
